+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ---
 ##5.9_WIP
 - Fixed -tv not working
+- Fixed Tweet vote notification showing incorrectly sometimes
+- Fixed error in ShowTweet
 
 ##5.8
 - Added -distance to get the direct distance between two cities